fastapi-client/pyproject.toml
[project]
name = "fastapi-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "anthropic>=0.77.0",
    "fastapi>=0.128.0",
    "fastapi-cache2>=0.2.2",
    "google-cloud-firestore>=2.23.0",
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "uvicorn>=0.40.0",
]
